centralize skill exceptions and standardize logging
